+++ frontend/src/metabase/databases/components/DatabaseEngineField/types.ts
@@ -1,0 +1,4 @@
+export interface EngineOption {
+  name: string;
+  value: string;
+}
